Updates routes and added necessary redirects
