+++ components/DatasetDetails/DatasetAboutInfo.vue
@@ -113,7 +113,7 @@ export default {
      getProjectLink: function(associatedProject) {
       const sys = propOr(null, 'sys', associatedProject)
       const entryId = propOr(null, 'id', sys)
-      return entryId != null ? `/projects/${entryId}` : ''
+      return entryId != null ? `/about/projects/${entryId}` : ''
     },
     /**
      * Construct the sparc award number
